// front-angular/src/app/models/user.model.ts
export interface User {
    '@id'?: string;
    firstname?: string;
    lastname?: string;
    address?: string;
    email?: string;
    phone_number?: string;
    phoneNumber?: string;
    roles?: any[];
    password?: string;
    id?: string;
    apiToken?: string;
    created_at?: any;
    action?: string;
}